requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
cmudict==1.0.32
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
joblib==1.4.2
lxml==5.4.0
numpy==2.2.5
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
PyPDF2==3.0.1
pyphen==0.17.2
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==79.0.1
sqlparse==0.5.3
textstat==0.7.5
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
zipp==3.21.0
